Update README.md
Edited the one file present to learn about branching and committing. I was unable to create a branch, per the direction provided by GitHub Hello-World example. However, while writing this edit I realized that there is an option below to create a new branch or commit the edit directly. Will choose to create a new-branch.